# Refactor Docker setup for smaller/ faster builds