--- a/.dockerignore
+++ b/.dockerignore
@@ -1,0 +1,5 @@
+__pycache__/
+.venv
+.vscode
+.pytest_cache
+tests/