--- a/Testing.Dockerfile
+++ b/Testing.Dockerfile
@@ -1,20 +1,24 @@
+# syntax=docker/dockerfile:1.3
+
 # This name comes from the docker-compose yml file that defines a name
 # for the "web" container's image.
 FROM kinto:build
 
-ENV PYTHONUNBUFFERED=1
+WORKDIR /app
+
+ENV PYTHONUNBUFFERED=1 \
+    PYTHONPATH="/app:$PYTHONPATH"
 
 USER root
 
 RUN apt-get update && \
     apt-get install -y --no-install-recommends \
-        curl wget
+    curl wget
 
 RUN pip install httpie kinto-http kinto-wizard
 
-WORKDIR /app
-ADD tests /app
-ADD kinto_remote_settings /app/kinto_remote_settings
+COPY tests .
+COPY kinto_remote_settings ./kinto_remote_settings
 
 ENTRYPOINT ["/bin/bash", "/app/run.sh"]
 CMD ["start"]