--- a/Testing.Dockerfile.dockerignore
+++ b/Testing.Dockerfile.dockerignore
@@ -1,0 +1,5 @@
+__pycache__/
+.venv
+.vscode
+.pytest_cache
+