--- a/bin/build-images.sh
+++ b/bin/build-images.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -eo pipefail
+
+docker build . -t kinto:build
+docker build . --file Testing.Dockerfile -t kinto:tests