… bin/docker-install.sh