- use multi-stage build process. Only download build dependencies in intermediate stage
- add `.dockerignore` for dist/test images
- reorder build steps for better layer caching

Also adds build script to build and tag dist, test images.

Size comparison of builds. Images with `-before` were built with code present in `main` (prior to this PR):
```
kinto               build          71db14f4f99a   5 minutes ago    349MB
kinto               tests          0debcb48417c   35 minutes ago   371MB
kinto               tests-before   2f90fb699378   35 minutes ago   1.4GB
kinto               build-before   8de540b3e6b2   37 minutes ago   1.37GB
```

Also, compare build times between the test suite run on the main branch [before](https://app.circleci.com/pipelines/github/mozilla-services/kinto-dist/1630/workflows/d4a0f208-a32b-45d2-bdac-26f3e650de5f) this PR and [after](https://app.circleci.com/pipelines/github/mozilla-services/kinto-dist/1634/workflows/53a2c0b7-ea9e-4af5-b3f7-8ba34ec68d9d).

Before:
```
build_test_publish   4m 37s
integration test     3m 2s
```

After:
```
build_test_publish   1m 59s
integration test     1m 49s
```